Criadas as tabelas do banco e separado arquivo de conexão do knex
